requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.8
packaging==24.1
paypalrestsdk==1.13.3
pillow==10.4.0
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dotenv==1.0.1
requests==2.32.3
rest-framework-simplejwt==0.0.2
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
whitenoise==6.7.0
